requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
dj-rest-auth==2.2.2
Django==4.0.1
django-allauth==0.47.0
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extensions==3.1.5
django-heroku==0.3.1
django-import-export==2.7.1
django-js-asset==1.2.2
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
MarkupPy==1.14
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tablib==3.2.0
urllib3==1.26.8
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
